Update current quarter logic
- Add setting (postLastClassDays) to ApiSettings that allows
specification of days after the year quarter last class day
to include quarter in getRegistrationQuarters
- Update App.config and Web.config transforms for nuget to
include this new config item.
- Update getRegistrationQuarters logic - it will now pull quarters
for which the registration date has passed and the last class day is
greater than postLastClassDays ago.
- The CurrentYearQuarter property does not use WebRegistrationSetting
data so its logic will remain slightly different. It pulls quarters for
which the LastClassDay has not passed then orders by quarter ID and takes
the first one. So, while the original query could return many future quarters,
ordering (ascending) by ID and then taking the first one will have the effect
of getting the current quarter
- Update to use just a Date rather than DateTime object in quarter queries.
YearQuarter data for LastClassDay and FirstRegistrationDate do not utilize
specific times. Thus it's safe to use a Date (which has a default time of
midnight 00:00:00) AND this means queries that use the same date will
actually be cached (whereas before because of DateTime the seconds
changed every query and essentially meant it didn't cache as expected).
- Addresses Jira issue CTCAPI-3